Migrate WatchVideoDescription to the composition API
